server/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
ecdsa==0.13.2
Flask==1.1.1
Flask-Cors==3.0.8
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pyasn1==0.4.7
pymongo==3.9.0
python-dotenv==0.10.3
python-jose==3.0.1
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.7
Werkzeug==0.16.0
